requirements.txt
pre-commit==3.6.0
black==23.12.1
faiss-cpu
#flake8==6.1.0
gigachat
python-dotenv
langchain
langchain-community
langchain-core
sentence-transformers
